# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Restaurant POS API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List orders
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [open, preparing, ready, closed, cancelled]
        - name: tableId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Create a new order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get an order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          description: Not found
    patch:
      operationId: updateOrder
      tags: [orders]
      summary: Update an order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{id}/items:
    post:
      operationId: addOrderItem
      tags: [orders]
      summary: Add item to an order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderItemInput"
      responses:
        "201":
          description: Item added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderItem"
  /orders/{id}/items/{itemId}:
    delete:
      operationId: removeOrderItem
      tags: [orders]
      summary: Remove an item from an order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Removed
  /orders/{id}/send-to-kitchen:
    post:
      operationId: sendOrderToKitchen
      tags: [orders, kitchen]
      summary: Send order to kitchen
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Sent to kitchen
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{id}/cancel:
    post:
      operationId: cancelOrder
      tags: [orders]
      summary: Cancel an order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Cancelled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"

  # Payments
  /payments:
    post:
      operationId: createPayment
      tags: [payments]
      summary: Process payment for an order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentInput"
      responses:
        "201":
          description: Payment processed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
  /payments/{orderId}/receipt:
    get:
      operationId: getReceipt
      tags: [payments]
      summary: Get receipt for an order
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Receipt data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Receipt"

  # Kitchen queue
...
components:
  schemas:
...
    Order:
      type: object
      required: [id, status, totalAmount, createdAt, items]
      properties:
        id:
          type: integer
        tableId:
          type: ["integer", "null"]
        tableNumber:
          type: ["integer", "null"]
        customerId:
          type: ["integer", "null"]
        customerName:
          type: ["string", "null"]
        status:
          type: string
          enum: [open, preparing, ready, closed, cancelled]
        type:
          type: string
          enum: [table, counter, takeaway]
        notes:
          type: ["string", "null"]
        totalAmount:
          type: number
        createdAt:
          type: string
        updatedAt:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
    OrderInput:
      type: object
      required: [type]
      properties:
        tableId:
          type: integer
        customerId:
          type: integer
        type:
          type: string
          enum: [table, counter, takeaway]
        notes:
          type: string
    OrderUpdate:
      type: object
      properties:
        tableId:
          type: integer
        customerId:
          type: integer
        status:
          type: string
          enum: [open, preparing, ready, closed, cancelled]
        notes:
          type: string
    OrderItem:
      type: object
      required: [id, orderId, productId, productName, quantity, unitPrice, totalPrice]
      properties:
        id:
          type: integer
        orderId:
          type: integer
        productId:
          type: integer
        productName:
          type: string
        quantity:
          type: integer
        unitPrice:
          type: number
        totalPrice:
          type: number
        notes:
          type: ["string", "null"]
    OrderItemInput:
      type: object
      required: [productId, quantity]
      properties:
        productId:
          type: integer
        quantity:
          type: integer
          minimum: 1
        notes:
          type: string

    # Payments
    Payment:
      type: object
      required: [id, orderId, amount, method, status, createdAt]
      properties:
        id:
          type: integer
        orderId:
          type: integer
        amount:
          type: number
        method:
          type: string
          enum: [cash, credit_card, debit_card, pix, voucher]
        status:
          type: string
          enum: [pending, approved, refused]
        change:
          type: ["number", "null"]
        createdAt:
          type: string
    PaymentInput:
      type: object
      required: [orderId, amount, method]
      properties:
        orderId:
          type: integer
        amount:
          type: number
          minimum: 0
        method:
          type: string
          enum: [cash, credit_card, debit_card, pix, voucher]
        amountTendered:
          type: number
    Receipt:
      type: object
      required: [order, payment, items]
      properties:
        order:
          $ref: "#/components/schemas/Order"
        payment:
          $ref: "#/components/schemas/Payment"
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"

    # Kitchen
...