# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Restaurant POS API specification
 * OpenAPI spec version: 0.1.0
 */
import type { OrderItem } from './orderItem';
import type { OrderStatus } from './orderStatus';
import type { OrderType } from './orderType';

export interface Order {
  id: number;
  /** @nullable */
  tableId?: number | null;
  /** @nullable */
  tableNumber?: number | null;
  /** @nullable */
  customerId?: number | null;
  /** @nullable */
  customerName?: string | null;
  status: OrderStatus;
  type?: OrderType;
  /** @nullable */
  notes?: string | null;
  totalAmount: number;
  createdAt: string;
  updatedAt?: string;
  items: OrderItem[];
}